+++ kernel/include/kernel/gdt.h
@@ -28,6 +28,6 @@ struct gdt_ptr {
 typedef struct gdt_ptr gdt_ptr_t;
 
 void gdt_bootstrap();
-void gdt_create_desc(int num, uint32_t base, uint32_t limit, uint16_t access, uint8_t gran);
+void gdt_create_desc(int num, uint32_t base, uint32_t limit, uint8_t access, uint8_t gran);
 
 #endif
